+++ dflib-csv/src/main/java/org/dflib/csv/ColumnConfig.java
@@ -13,18 +13,18 @@ import java.util.function.IntFunction;
 
 class ColumnConfig {
 
-    int csvColPos;
-    String csvColName;
+    int srcColPos;
+    String srcColName;
     IntFunction<Extractor<CSVRecord, ?>> extractorMaker;
     boolean compact;
 
     private ColumnConfig() {
-        csvColPos = -1;
+        srcColPos = -1;
     }
 
     public static ColumnConfig objectCol(int pos, boolean compact) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColPos = pos;
+        config.srcColPos = pos;
         config.extractorMaker = i -> Extractor.$col(r -> r.get(i));
         config.compact = compact;
         return config;
@@ -32,7 +32,7 @@ class ColumnConfig {
 
     public static ColumnConfig objectCol(String name, boolean compact) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColName = name;
+        config.srcColName = name;
         config.extractorMaker = i -> Extractor.$col(r -> r.get(i));
         config.compact = compact;
         return config;
@@ -40,7 +40,7 @@ class ColumnConfig {
 
     public static ColumnConfig objectCol(int pos, ValueMapper<String, ?> mapper, boolean compact) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColPos = pos;
+        config.srcColPos = pos;
         config.extractorMaker = i -> Extractor.$col(r -> mapper.map(r.get(i)));
         config.compact = compact;
         return config;
@@ -48,7 +48,7 @@ class ColumnConfig {
 
     public static ColumnConfig objectCol(String name, ValueMapper<String, ?> mapper, boolean compact) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColName = name;
+        config.srcColName = name;
         config.extractorMaker = i -> Extractor.$col(r -> mapper.map(r.get(i)));
         config.compact = compact;
         return config;
@@ -56,62 +56,62 @@ class ColumnConfig {
 
     public static ColumnConfig intCol(int pos, IntValueMapper<String> mapper) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColPos = pos;
+        config.srcColPos = pos;
         config.extractorMaker = i -> Extractor.$int(r -> mapper.map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig intCol(String name, IntValueMapper<String> mapper) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColName = name;
+        config.srcColName = name;
         config.extractorMaker = i -> Extractor.$int(r -> mapper.map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig longCol(int pos, LongValueMapper<String> mapper) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColPos = pos;
+        config.srcColPos = pos;
         config.extractorMaker = i -> Extractor.$long(r -> mapper.map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig longCol(String name, LongValueMapper<String> mapper) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColName = name;
+        config.srcColName = name;
         config.extractorMaker = i -> Extractor.$long(r -> mapper.map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig doubleCol(int pos, DoubleValueMapper<String> mapper) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColPos = pos;
+        config.srcColPos = pos;
         config.extractorMaker = i -> Extractor.$double(r -> mapper.map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig doubleCol(String name, DoubleValueMapper<String> mapper) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColName = name;
+        config.srcColName = name;
         config.extractorMaker = i -> Extractor.$double(r -> mapper.map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig boolCol(int pos) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColPos = pos;
+        config.srcColPos = pos;
         config.extractorMaker = i -> Extractor.$bool(r -> BoolValueMapper.fromString().map(r.get(i)));
         return config;
     }
 
     public static ColumnConfig boolCol(String name) {
         ColumnConfig config = new ColumnConfig();
-        config.csvColName = name;
+        config.srcColName = name;
         config.extractorMaker = i -> Extractor.$bool(r -> BoolValueMapper.fromString().map(r.get(i)));
         return config;
     }
 
     public Extractor<CSVRecord, ?> extractor(Index csvHeader) {
-        int csvPos = csvColPos >= 0 ? csvColPos : csvHeader.position(csvColName);
+        int csvPos = srcColPos >= 0 ? srcColPos : csvHeader.position(srcColName);
         Extractor<CSVRecord, ?> e = extractorMaker.apply(csvPos);
         return compact ? e.compact() : e;
     }